Add support for doctrine/persistence 4
v4 provides a guarantee that ManagerRegistry::getManager() returns an entity
manager (as opposed to null).

The tests need to be adjusted to reflect the behavior of the mocked dependency
more accurately, as it is throwing an exception in case of a null manager for
all three supported versions of the library.