skip interactive questions asked by Composer
following #60587 so that the installation script is not blocked by
Composer asking to install the bridge as a dev dependency:

```
The package you required is recommended to be placed in require-dev (because it is tagged as "testing") but you did not use --dev.
Do you want to re-run the command with --dev? [yes]?
```