Fix TraceableSerializer when collected caller from array map
If the TraceableSerializer runs from a callable in an array_map, then
the caller looses the file and the line because it was invoked. This
causes a hard fail since the required items are not defined. The error
is the following:
TraceableSerializer.php line 174:
Warning: Undefined array key "file".

The fix is to get the file and the line from the following array item
which always is the caller class.